+++ src/spade/__init__.py
@@ -1,6 +1,6 @@
 """SPADE - Sub-Patch Analysis & Detection Engine"""
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 
 from spade.engine import ForensicsEngine, Config